# bos-workspace/backend-nestjs/openapi-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/entities": {
      "post": {
        "operationId": "EntitiesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEntityDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Tạo Entity mới (VD: Phiếu xin nghỉ)",
        "tags": [
          "Metadata - Entities (Thực thể / Biểu mẫu)"
        ]
      },
      "get": {
        "operationId": "EntitiesController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "schema": {
              "example": "id",
              "type": "string"
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "asc",
                "desc"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Lấy danh sách các Entities (Có phân trang & Sắp xếp)",
        "tags": [
          "Metadata - Entities (Thực thể / Biểu mẫu)"
        ]
      }
    },
    "/api/v1/entities/{id}": {
      "get": {
        "operationId": "EntitiesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Lấy chi tiết Entity kèm danh sách các trường (Fields)",
        "tags": [
          "Metadata - Entities (Thực thể / Biểu mẫu)"
        ]
      },
      "patch": {
        "operationId": "EntitiesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEntityDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Cập nhật thông tin Entity",
        "tags": [
          "Metadata - Entities (Thực thể / Biểu mẫu)"
        ]
      },
      "delete": {
        "operationId": "EntitiesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Xóa Entity",
        "tags": [
          "Metadata - Entities (Thực thể / Biểu mẫu)"
        ]
      }
    },
    "/api/v1/entities/{id}/versions": {
      "get": {
        "operationId": "EntitiesController_findVersions",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Lấy lịch sử phiên bản của Entity",
        "tags": [
          "Metadata - Entities (Thực thể / Biểu mẫu)"
        ]
      }
    },
    "/api/v1/entities/{id}/versions/{versionId}/restore": {
      "post": {
        "operationId": "EntitiesController_restoreVersion",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "versionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Khôi phục Entity về một phiên bản cũ",
        "tags": [
          "Metadata - Entities (Thực thể / Biểu mẫu)"
        ]
      }
    },
...
    "/api/v1/print-templates": {
      "post": {
        "operationId": "PrintTemplatesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTemplateDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Tạo mới một mẫu in cho Biểu mẫu",
        "tags": [
          "Metadata - Print Templates (Mẫu in động)"
        ]
      },
      "get": {
        "operationId": "PrintTemplatesController_findAll",
        "parameters": [
          {
            "name": "entityId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Lấy danh sách tất cả mẫu in của Biểu mẫu",
        "tags": [
          "Metadata - Print Templates (Mẫu in động)"
        ]
      }
    },
    "/api/v1/print-templates/{id}": {
      "get": {
        "operationId": "PrintTemplatesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Lấy chi tiết cấu hình mẫu in",
        "tags": [
          "Metadata - Print Templates (Mẫu in động)"
        ]
      },
      "patch": {
        "operationId": "PrintTemplatesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTemplateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Cập nhật mẫu in",
        "tags": [
          "Metadata - Print Templates (Mẫu in động)"
        ]
      },
      "delete": {
        "operationId": "PrintTemplatesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Xóa mẫu in",
        "tags": [
          "Metadata - Print Templates (Mẫu in động)"
        ]
      }
    },
    "/api/v1/print-templates/{id}/render/{recordId}": {
      "get": {
        "operationId": "PrintTemplatesController_renderTemplate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "recordId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Biên dịch dữ liệu bản ghi vào Mẫu in ra mã HTML hoàn chỉnh",
        "tags": [
          "Metadata - Print Templates (Mẫu in động)"
        ]
      }
    },
...
    "/api/v1/tasks/my-tasks": {
      "get": {
        "operationId": "TasksController_findMyTasks",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "example": "PENDING",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Lấy danh sách nhiệm vụ của tôi (Có phân trang & lọc trạng thái)",
        "tags": [
          "Work Engine - Tasks (Nhiệm vụ & Giao việc)"
        ]
      }
    },
    "/api/v1/tasks/analytics": {
      "get": {
        "operationId": "TasksController_getAnalytics",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Thống kê KPI xử lý nhiệm vụ toàn công ty (Dành cho Quản lý)",
        "tags": [
          "Work Engine - Tasks (Nhiệm vụ & Giao việc)"
        ]
      }
    },
    "/api/v1/tasks/{id}/complete": {
      "post": {
        "operationId": "TasksController_completeTask",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompleteTaskDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Hoàn thành nhiệm vụ (Tự động tính thời gian SLA)",
        "tags": [
          "Work Engine - Tasks (Nhiệm vụ & Giao việc)"
        ]
      }
    },
    "/api/v1/tasks/batch-complete": {
      "post": {
        "operationId": "TasksController_batchCompleteTasks",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchCompleteTasksDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Duyệt nhanh hàng loạt nhiệm vụ",
        "tags": [
          "Work Engine - Tasks (Nhiệm vụ & Giao việc)"
        ]
      }
    },
    "/api/v1/tasks/{id}/delegate": {
      "post": {
        "operationId": "TasksController_delegateTask",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DelegateTaskDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Ủy quyền nhiệm vụ cho người khác",
        "tags": [
          "Work Engine - Tasks (Nhiệm vụ & Giao việc)"
        ]
      }
    },
    "/api/v1/tasks": {
      "get": {
        "operationId": "TasksController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Lấy toàn bộ danh sách Task (Super Admin)",
        "tags": [
          "Work Engine - Tasks (Nhiệm vụ & Giao việc)"
        ]
      }
    },
...
  },
  "info": {
    "title": "BOS Core Engine API",
    "description": "Tài liệu API cho Nền tảng Quản trị Doanh nghiệp (BOS v2.0)",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateEntityDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Tên hiển thị của Entity",
            "example": "Phiếu đề xuất mua sắm"
          },
          "code": {
            "type": "string",
            "description": "Mã code duy nhất, viết hoa không dấu",
            "example": "PURCHASE_REQUEST"
          },
          "description": {
            "type": "string",
            "description": "Mô tả chi tiết"
          },
          "autoCodePattern": {
            "type": "string",
            "description": "Mẫu tự sinh mã nghiệp vụ (Hỗ trợ cấu hình dạng SEQ)",
            "example": "QTMS-{SEQ:4}"
          },
          "titlePattern": {
            "type": "string",
            "description": "Mẫu tiêu đề hồ sơ động",
            "example": "Đề xuất Thanh toán - {supplier_name} - {RECORD_CODE}"
          }
        },
        "required": [
          "name",
          "code"
        ]
      },
      "UpdateEntityDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Tên hiển thị của Entity",
            "example": "Phiếu đề xuất mua sắm"
          },
          "code": {
            "type": "string",
            "description": "Mã code duy nhất, viết hoa không dấu",
            "example": "PURCHASE_REQUEST"
          },
          "description": {
            "type": "string",
            "description": "Mô tả chi tiết"
          },
          "autoCodePattern": {
            "type": "string",
            "description": "Mẫu tự sinh mã nghiệp vụ (Hỗ trợ cấu hình dạng SEQ)",
            "example": "QTMS-{SEQ:4}"
          },
          "titlePattern": {
            "type": "string",
            "description": "Mẫu tiêu đề hồ sơ động",
            "example": "Đề xuất Thanh toán - {supplier_name} - {RECORD_CODE}"
          }
        }
      },
...
      "CreateTemplateDto": {
        "type": "object",
        "properties": {
          "entityId": {
            "type": "number",
            "description": "ID của Entity (Biểu mẫu)",
            "example": 1
          },
          "name": {
            "type": "string",
            "description": "Tên mẫu in",
            "example": "Mẫu quyết định mua sắm trang thiết bị"
          },
          "template": {
            "type": "object",
            "description": "Cấu hình mẫu in chứa HTML (Lưu trong Json)",
            "example": {
              "html": "<h1>PHIẾU ĐỀ XUẤT MUA SẮM</h1><p>Lý do: {{data.reason}}</p>"
            }
          }
        },
        "required": [
          "entityId",
          "name",
          "template"
        ]
      },
      "UpdateTemplateDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Tên mẫu in",
            "example": "Mẫu quyết định mua sắm trang thiết bị"
          },
          "template": {
            "type": "object",
            "description": "Cấu hình mẫu in chứa HTML (Lưu trong Json)",
            "example": {
              "html": "<h1>PHIẾU ĐỀ XUẤT MUA SẮM</h1><p>Lý do: {{data.reason}}</p>"
            }
          }
        }
      },
...
      "CompleteTaskDto": {
        "type": "object",
        "properties": {
          "comment": {
            "type": "string",
            "description": "Ý kiến khi hoàn thành/xử lý nhiệm vụ"
          }
        }
      },
      "BatchCompleteTasksDto": {
        "type": "object",
        "properties": {
          "taskIds": {
            "description": "Danh sách ID nhiệm vụ cần phê duyệt",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "comment": {
            "type": "string",
            "description": "Ý kiến phê duyệt chung"
          }
        }
      },
      "DelegateTaskDto": {
        "type": "object",
        "properties": {
          "assigneeId": {
            "type": "number",
            "description": "ID người được ủy quyền"
          },
          "reason": {
            "type": "string",
            "description": "Lý do ủy quyền"
          }
        }
      },
...
    }
  }
}